requirements.txt
APScheduler
click
psycopg2
python-dateutil
pytz
redis
requests
rq
six
slackclient
SQLAlchemy
tzlocal
websocket-client
beautifulsoup4
bs4
html5lib
